chore(xm-rating): move to packages/
