+++ packages/components/src/xm-rating/index.ts
